genirq/msi: Optionally use dev->fwnode for device domain
mainline inclusion
commit 9d1c58c8004653b37721dd7b16f4360216778c94
category: feature
bugzilla: https://github.com/RVCK-Project/rvck-olk/issues/1

--------------------------------

To support wire to MSI domains via the MSI infrastructure it is required to
use the firmware node of the device which implements this for creating the
MSI domain. Otherwise the existing firmware match mechanisms to find the
correct irqdomain for a wired interrupt which is connected to a wire to MSI
bridge would fail.

This cannot be used for the general case because not all devices provide
firmware nodes and all regular per device MSI domains are directly
associated to the device and have not be searched for.

Signed-off-by: Thomas Gleixner <tglx@linutronix.de>
Signed-off-by: Anup Patel <apatel@ventanamicro.com>
Signed-off-by: Thomas Gleixner <tglx@linutronix.de>
Link: https://lore.kernel.org/r/20240127161753.114685-11-apatel@ventanamicro.com
Signed-off-by: Hangfan Li <lihangfan@iscas.ac.cn>